1016: Rename SVG templates to .svgt to stop confusing GitHub.
