+++ accessibility_monitoring_platform/apps/common/templates/common/metrics/chart.svgt
